+++ pkg/api/schema/convert.go
@@ -1,3 +1,19 @@
+/*
+Copyright 2019-2020 vChain, Inc.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+	http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
 package schema
 
 import (
@@ -25,10 +41,7 @@ func (item *Item) ToSItem() (*StructuredItem, error) {
 	return &StructuredItem{
 		Index: item.Index,
 		Key:   item.Key,
-		Value: &Content{
-			Payload:   c.Payload,
-			Timestamp: c.Timestamp,
-		},
+		Value: &c,
 	}, nil
 }
